requirements.txt
aiogram
gigachain
langchain_community